+++ libs/vue/guard/metamask.guard.ts
@@ -1,24 +1,24 @@
 import Vue from 'vue';
 import { NavigationGuardNext, Route } from 'vue-router';
-import { getEthersService } from '../plugin/ethers.plugin';
+import { getMetamaskService } from '../plugin/metamask.plugin';
 
-export function EthersGuard(
+export function MetamaskGuard(
   _to: Route,
   _from: Route,
   next: NavigationGuardNext<Vue>,
 ): void {
-  const ethersService = getEthersService();
+  const metamaskService = getMetamaskService();
 
   const fn = (): void => {
-    if (ethersService.isAuthenticated) return next();
+    if (metamaskService.isAuthenticated) return next();
     return next({ name: 'index' });
   };
 
-  ethersService.$watch('loading', (loading) => {
+  metamaskService.$watch('loading', (loading) => {
     if (!loading) fn();
     return undefined;
   });
 
-  if (!ethersService.loading) return fn();
+  if (!metamaskService.loading) return fn();
   return undefined;
 }